chatbotapp/pyproject.toml
[tool.poetry]
name = "kintsugi-tax-bot-master"
version = "0.1.0"
description = ""
authors = ["kintsugi-tax-bot"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
openai = "^1.6.1"
langchain = "^0.0.352"
tiktoken = "0.5.2"
pgvector = "0.2.4"
psycopg2-binary = "2.9.9"
uvicorn = "0.25.0"
jinja2 = "3.1.2"
python-multipart = "^0.0.6"
pytest = "^7.4.3"
llama-index = "0.9.22"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
deepeval = "^0.20.44"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
